New validator core dependency (#63)
* Add version selection, update validator explanation

The validator may validate against versions 3.0, 2.2, and 2.1.
This commit is a Work In Progress: take note of the local dependency
path for the hpt-validator module. This change must not be merged in
until an updated version of the hpt-validator of the module is available
and these changes have been tested with that module.

* version validator bump

* updated with npm audit fixes

* optionalDependencies flag added

---------

Co-authored-by: Mint Thompson <mathompson@mitre.org>